+++ usr/src/grub/grub-0.97/stage2/stage1_5.c
@@ -16,13 +16,16 @@
  *  along with this program; if not, write to the Free Software
  *  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
+/*
+ * Copyright 2016 Nexenta Systems, Inc.
+ */
 
 #include "shared.h"
 
-static unsigned int saved_sector = (unsigned int)-1;
+static unsigned long long saved_sector = (unsigned long long)-1;
 
 static void
-disk_read_savesect_func (unsigned int sector, int offset, int length)
+disk_read_savesect_func (unsigned long long sector, int offset, int length)
 {
   saved_sector = sector;
 }
